+++ src/pages/ProductListing/ProductListing.css
@@ -31,9 +31,10 @@
 }
 
 .grid-product-layout .card .card-subtitle {
-	width: 14rem;
+	width: 13rem;
 	overflow: hidden;
-	white-space: wrap;
+	white-space: nowrap;
+	text-overflow: ellipsis;
 }
 
 .gap-xs {
