+++ src/services/cart-services/index.js
@@ -1,0 +1,4 @@
+export { addToCartService } from "./addToCartService";
+export { cartCounterService } from "./cartCounterService";
+export { getCartService } from "./getCartService";
+export { removeProductCartService } from "./removeProductCartService";
