+++ src/services/wishlist-services/index.js
@@ -1,0 +1,3 @@
+export { addToWishlistService } from "./addToWishlistService";
+export { getWishlistService } from "./getWishlistService";
+export { removeProductWishlistService } from "./removeProductWishlistService";
